refactor(workflow): remove CheckSkew step and related configurations
