+++ ui/src/hooks/useAnalysisChat.ts
@@ -3,6 +3,7 @@
 import { useState, useCallback, useRef, useEffect } from "react";
 import { flushSync } from "react-dom";
 
+import { buildAuthHeaders } from "@/lib/auth/session";
 import type { ModelOverride } from "@/lib/copilotModels";
 
 export interface ChatMessage {
@@ -54,42 +55,8 @@ export interface AnalysisContext {
   taskId: string;
 }
 
-const PROJECT_STORAGE_KEY = "qdash_current_project_id";
-
 function buildHeaders(): Record<string, string> {
-  const headers: Record<string, string> = {
-    "Content-Type": "application/json",
-  };
-
-  // Match custom-instance.ts interceptor: cookie "access_token"
-  const accessToken = document.cookie
-    .split("; ")
-    .find((row) => row.startsWith("access_token="))
-    ?.split("=")[1];
-  if (accessToken) {
-    headers["Authorization"] = `Bearer ${decodeURIComponent(accessToken)}`;
-  }
-
-  // Match AxiosProvider interceptor: cookie "token" (username)
-  const token = document.cookie
-    .split("; ")
-    .find((row) => row.startsWith("token="))
-    ?.split("=")[1];
-  if (token) {
-    const decoded = decodeURIComponent(token);
-    if (!headers["Authorization"]) {
-      headers["Authorization"] = `Bearer ${decoded}`;
-    }
-    headers["X-Username"] = decoded;
-  }
-
-  // Match AxiosProvider: localStorage project ID
-  const projectId = localStorage.getItem(PROJECT_STORAGE_KEY);
-  if (projectId) {
-    headers["X-Project-Id"] = projectId;
-  }
-
-  return headers;
+  return buildAuthHeaders();
 }
 
 interface SSEEvent {
